fix ts-ignore errors refactored large functions to use helpers
