+++ packages/mermaid/src/diagrams/git/gitGraphTypes.ts
@@ -1,4 +1,13 @@
-export type CommitType = 'NORMAL' | 'REVERSE' | 'HIGHLIGHT' | 'MERGE' | 'CHERRY_PICK';
+import type { DiagramDB } from '../../diagram-api/types.js';
+import type { GitGraphDiagramConfig } from '../../config.type.js';
+
+export interface CommitType {
+  NORMAL: number;
+  REVERSE: number;
+  HIGHLIGHT: number;
+  MERGE: number;
+  CHERRY_PICK: number;
+}
 
 export interface Commit {
   id: string;
@@ -14,6 +23,11 @@ export interface Commit {
 
 export interface GitGraph {
   statements: Statement[];
+}
+
+export interface Position {
+  x: number;
+  y: number;
 }
 
 export type Statement = CommitAst | BranchAst | MergeAst | CheckoutAst | CherryPickingAst;
@@ -50,6 +64,49 @@ export interface CherryPickingAst {
   id: string;
   tags?: string[];
   parent: string;
+}
+
+export interface GitGraphDB extends DiagramDB {
+  // config
+  getConfig: () => GitGraphDiagramConfig | undefined;
+
+  // common db
+  clear: () => void;
+  setDiagramTitle: (title: string) => void;
+  getDiagramTitle: () => string;
+  setAccTitle: (title: string) => void;
+  getAccTitle: () => string;
+  setAccDescription: (description: string) => void;
+  getAccDescription: () => string;
+
+  // diagram db
+  commitType: CommitType;
+  setDirection: (direction: DiagramOrientation) => void;
+  getDirection: () => DiagramOrientation;
+  setOptions: (options: string) => void;
+  getOptions: () => string;
+  commit: (msg: string, id: string, type: number, tags?: string[] | undefined) => void;
+  branch: (name: string, order: number) => void;
+  merge: (
+    otherBranch: string,
+    customId?: string,
+    overrideType?: number,
+    customTags?: string[]
+  ) => void;
+  cherryPick: (
+    sourceId: string,
+    targetId: string,
+    tags: string[] | undefined,
+    parentCommitId: string
+  ) => void;
+  checkout: (branch: string) => void;
+  prettyPrint: () => void;
+  getBranchesAsObjArray: () => { name: string }[];
+  getBranches: () => Map<string, string | null>;
+  getCommits: () => Map<string, Commit>;
+  getCommitsArray: () => Commit[];
+  getCurrentBranch: () => string;
+  getHead: () => Commit | null;
 }
 
 export type DiagramOrientation = 'LR' | 'TB';
